use site date and time formattings #16

+++ source/includes/Settings.php
@@ -4,12 +4,11 @@ namespace MobilizonConnector;
 class Settings {
 
   private static $DEFAULT_OPTION_URL = 'https://mobilizon.fr';
-  private static $DEFAULT_IS_SHORT_OFFSET_NAME_SHOWN = false;
   private static $PAGE_NAME = 'wordpress_mobilizon';
   private static $OPTIONS_GROUP_NAME = 'wordpress_mobilizon';
   private static $OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN = 'wordpress_mobilizon_is_short_offset_name_shown';
+  private static $OPTION_NAME_PLUGIN_VERSION = 'wordpress_mobilizon_plugin_version';
   private static $OPTION_NAME_URL = 'wordpress_mobilizon_url';
-  private static $SETTING_FIELD_NAME_IS_SHORT_OFFSET_NAME_SHOWN = 'wordpress_mobilizon_field_is_short_offset_name_shown';
   private static $SETTING_FIELD_NAME_URL = 'wordpress_mobilizon_field_url';
   private static $SETTINGS_SECTION_NAME = 'wordpress_mobilizon_section_general';
 
@@ -19,10 +18,6 @@ class Settings {
   }
 
   public static function init_settings() {
-    register_setting(
-      self::$OPTIONS_GROUP_NAME,
-      self::$OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN
-    );
     register_setting(
       self::$OPTIONS_GROUP_NAME,
       self::$OPTION_NAME_URL,
@@ -46,21 +41,6 @@ class Settings {
         'label_for' => self::$SETTING_FIELD_NAME_URL
       )
     );
-    add_settings_field(
-      self::$SETTING_FIELD_NAME_IS_SHORT_OFFSET_NAME_SHOWN,
-      esc_html__('Display named offset', 'connector-mobilizon'),
-      'MobilizonConnector\Settings::output_field_is_short_offset_name_shown',
-      self::$PAGE_NAME,
-      self::$SETTINGS_SECTION_NAME,
-      array(
-        'label_for' => self::$SETTING_FIELD_NAME_IS_SHORT_OFFSET_NAME_SHOWN
-      )
-    );
-  }
-
-  public static function output_field_is_short_offset_name_shown($args) {
-    $isShortOffsetNameShown = self::isShortOffsetNameShown();
-    require dirname(__DIR__) . '/view/settings/is-short-offset-name-shown-field.php';
   }
 
   public static function output_field_url($args) {
@@ -101,22 +81,26 @@ class Settings {
     require dirname(__DIR__) . '/view/settings/page.php';
   }
 
-  public static function isShortOffsetNameShown() {
-    return get_option(self::$OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN);
-  }
-
   public static function getUrl() {
     return get_option(self::$OPTION_NAME_URL);
   }
 
   public static function setDefaultOptions() {
-    add_option(self::$OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN, self::$DEFAULT_IS_SHORT_OFFSET_NAME_SHOWN);
     add_option(self::$OPTION_NAME_URL, self::$DEFAULT_OPTION_URL);
   }
 
   public static function deleteAllOptions() {
-    delete_option(self::$OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN);
     delete_option(self::$OPTION_NAME_URL);
+  }
+
+  public static function removeObsoleteOptionsIfNeeded() {
+    $storedPluginVersion = get_option(self::$OPTION_NAME_PLUGIN_VERSION);
+    if ($storedPluginVersion !== PLUGIN_VERSION) {
+      if (version_compare($storedPluginVersion, '1.5.0', '<') ) {
+        delete_option(self::$OPTION_NAME_IS_SHORT_OFFSET_NAME_SHOWN);
+      }
+      update_option(self::$OPTION_NAME_PLUGIN_VERSION, PLUGIN_VERSION);
+    }    
   }
 
 }
